IS-1977: Refactor historikk folder into /sider folder
